@@ -3,28 +3,35 @@
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
 import errno
+import os
 from setuptools import setup, find_packages
 from setuptools.command.install import install
 from subprocess import check_call
 
 
-VERSION = "0.0.0"
+VERSION = os.getenv("PULUMI_PYTHON_VERSION", "0.0.0")
 def readme():
     try:
         with open('README.md', encoding='utf-8') as f:
             return f.read()
     except FileNotFoundError:
-        return "xyz Pulumi Package - Development Version"
+        return "docker-native Pulumi Package - Development Version"
 
 
-setup(name='pulumi_xyz',
-      python_requires='>=3.7',
+setup(name='pulumi_docker_native',
+      python_requires='>=3.8',
       version=VERSION,
+      description="Description",
       long_description=readme(),
       long_description_content_type='text/markdown',
+      keywords='keywords',
+      url='pulumi.com',
+      project_urls={
+          'Repository': 'https://github.com/pulumi/pulumi-docker-native'
+      },
       packages=find_packages(),
       package_data={
-          'pulumi_xyz': [
+          'pulumi_docker_native': [
               'py.typed',
               'pulumi-plugin.json',
           ]
