@@ -1,7 +1,7 @@
 // *** WARNING: this file was generated by pulumi. ***
 // *** Do not edit by hand unless you're certain you know what you are doing! ***
 
-namespace Pulumi.Xyz
+namespace Pulumi.DockerNative
 {
     static class Utilities
     {
@@ -53,6 +53,7 @@ namespace Pulumi.Xyz
         {
             var dst = src ?? new global::Pulumi.InvokeOptions{};
             dst.Version = src?.Version ?? Version;
+            dst.PluginDownloadURL = src?.PluginDownloadURL ?? "github.com/pulumi/pulumi-docker-native";
             return dst;
         }
 
@@ -62,7 +63,7 @@ namespace Pulumi.Xyz
         static Utilities()
         {
             var assembly = global::System.Reflection.IntrospectionExtensions.GetTypeInfo(typeof(Utilities)).Assembly;
-            using var stream = assembly.GetManifestResourceStream("Pulumi.Xyz.version.txt");
+            using var stream = assembly.GetManifestResourceStream("Pulumi.DockerNative.version.txt");
             using var reader = new global::System.IO.StreamReader(stream ?? throw new global::System.NotSupportedException("Missing embedded version.txt file"));
             version = reader.ReadToEnd().Trim();
             var parts = version.Split("\n");
@@ -74,9 +75,9 @@ namespace Pulumi.Xyz
         }
     }
 
-    internal sealed class XyzResourceTypeAttribute : global::Pulumi.ResourceTypeAttribute
+    internal sealed class DockerNativeResourceTypeAttribute : global::Pulumi.ResourceTypeAttribute
     {
-        public XyzResourceTypeAttribute(string type) : base(type, Utilities.Version)
+        public DockerNativeResourceTypeAttribute(string type) : base(type, Utilities.Version)
         {
         }
     }
