@@ -1,13 +1,13 @@
 // Code generated by pulumi-language-go DO NOT EDIT.
 // *** WARNING: Do not edit by hand unless you're certain you know what you are doing! ***
 
-package xyz
+package docker
 
 import (
 	"fmt"
 
 	"github.com/blang/semver"
-	"github.com/pulumi/pulumi-docker-native/sdk/go/xyz/internal"
+	"github.com/pulumi/pulumi-docker-native/sdk/go/internal"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
 
@@ -21,7 +21,7 @@ func (m *module) Version() semver.Version {
 
 func (m *module) Construct(ctx *pulumi.Context, name, typ, urn string) (r pulumi.Resource, err error) {
 	switch typ {
-	case "xyz:index:Random":
+	case "docker-native:index:Random":
 		r = &Random{}
 	default:
 		return nil, fmt.Errorf("unknown resource type: %s", typ)
@@ -40,7 +40,7 @@ func (p *pkg) Version() semver.Version {
 }
 
 func (p *pkg) ConstructProvider(ctx *pulumi.Context, name, typ, urn string) (pulumi.ProviderResource, error) {
-	if typ != "pulumi:providers:xyz" {
+	if typ != "pulumi:providers:docker-native" {
 		return nil, fmt.Errorf("unknown provider type: %s", typ)
 	}
 
@@ -55,12 +55,12 @@ func init() {
 		version = semver.Version{Major: 1}
 	}
 	pulumi.RegisterResourceModule(
-		"xyz",
+		"docker-native",
 		"index",
 		&module{version},
 	)
 	pulumi.RegisterResourcePackage(
-		"xyz",
+		"docker-native",
 		&pkg{version},
 	)
 }
