@@ -1,14 +1,15 @@
 // Code generated by pulumi-language-go DO NOT EDIT.
 // *** WARNING: Do not edit by hand unless you're certain you know what you are doing! ***
 
-package xyz
+package docker
 
 import (
 	"context"
 	"reflect"
 
-	"github.com/pulumi/pulumi-docker-native/sdk/go/xyz/internal"
+	"github.com/pulumi/pulumi-docker-native/sdk/go/internal"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+	"github.com/pulumi/pulumi/sdk/v3/go/pulumix"
 )
 
 type Provider struct {
@@ -24,7 +25,7 @@ func NewProvider(ctx *pulumi.Context,
 
 	opts = internal.PkgResourceDefaultOpts(opts)
 	var resource Provider
-	err := ctx.RegisterResource("pulumi:providers:xyz", name, args, &resource, opts...)
+	err := ctx.RegisterResource("pulumi:providers:docker-native", name, args, &resource, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -42,29 +43,10 @@ func (ProviderArgs) ElementType() reflect.Type {
 	return reflect.TypeOf((*providerArgs)(nil)).Elem()
 }
 
-type ProviderInput interface {
-	pulumi.Input
-
-	ToProviderOutput() ProviderOutput
-	ToProviderOutputWithContext(ctx context.Context) ProviderOutput
-}
-
-func (*Provider) ElementType() reflect.Type {
-	return reflect.TypeOf((**Provider)(nil)).Elem()
-}
-
-func (i *Provider) ToProviderOutput() ProviderOutput {
-	return i.ToProviderOutputWithContext(context.Background())
-}
-
-func (i *Provider) ToProviderOutputWithContext(ctx context.Context) ProviderOutput {
-	return pulumi.ToOutputWithContext(ctx, i).(ProviderOutput)
-}
-
 type ProviderOutput struct{ *pulumi.OutputState }
 
 func (ProviderOutput) ElementType() reflect.Type {
-	return reflect.TypeOf((**Provider)(nil)).Elem()
+	return reflect.TypeOf((*Provider)(nil)).Elem()
 }
 
 func (o ProviderOutput) ToProviderOutput() ProviderOutput {
@@ -75,7 +57,12 @@ func (o ProviderOutput) ToProviderOutputWithContext(ctx context.Context) Provide
 	return o
 }
 
+func (o ProviderOutput) ToOutput(ctx context.Context) pulumix.Output[Provider] {
+	return pulumix.Output[Provider]{
+		OutputState: o.OutputState,
+	}
+}
+
 func init() {
-	pulumi.RegisterInputType(reflect.TypeOf((*ProviderInput)(nil)).Elem(), &Provider{})
 	pulumi.RegisterOutputType(ProviderOutput{})
 }
